Ignore passed consumers in allocation candidates
This commit makes it possible to pass a list of consumer UUIDs to ignore
when retrieven allocation candidates. This can be useful for
implementations replacing a VM on resize instead of copying it. Some
care has to be taken on the client side to not overbook the
resource-provider, but there's another check when trying to allocate the
resources that should make sure of that.