Allow switching resources on overused providers
When a resource-provider gets overused, e.g. because a host in the cluster it
represents goes down or because one memory DIMM of multiple fails, we need to
be able to (offline) migrate/resize a VM to reduce the pressure on the host.

For that, we check in the call setting the allocations if for every
allocation with a value for a resource, there's also an allocation
removing such a resource - namely switching the allocations from one
consumer to another as is done for migrations.

Change-Id: Icb985502bd6277dc8ab5fc6602dde501d07eef65