tests: Never run with "independent" transaction
With the tests running on sqlite in-memory DBs, they cannot properly
support independent transactions. This comes from the fact that every
transaction started like that is either committed or rolled back - and
with it all the other data of the other transactions in the same
connection. To properly support this, "independent" would have to create
a new connection instead of only a transaction in the same connection.

Since we do not have a test-case relying on "independent" but just code
that uses it, we make "independent" a NOOP by monkey-patching the
`_clone()` method of `oslo.db`'s `_TransactionContextManager` to always
set the `independent` parameter to `False`.

Change-Id: I24b0446f81634bbb6b7ad672dd51a572043f72af