Exclude own consumers in capacity check
Since Ia2d57e0f8b963a00d2f26c447f1d2fb0a108555e we run the capacity
check in its own transaction. Reading outside of our transaction, the
deleting of allocations of the consumers we currently handle is not
visible. Therefore, we have to exclude the resources from our consumers
from the sum of resources of the resource providers we handle.
Otherwise, we would count the old and the new resources of our consumers
in the capacity check and would thus report exceeded capacity where
there is none.

Change-Id: I287e8167aa61ce9931621fbe836a97d9bfc668d1